# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RT Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface KartuKeluarga {
  id: number;
  nomorKK: string;
  nomorRumah: string;
  namaKepalaKeluarga: string;
  createdAt: string;
  /** @nullable */
  wargaCount?: number | null;
}

export interface KartuKeluargaInput {
  nomorKK: string;
  nomorRumah: string;
  namaKepalaKeluarga: string;
}

export interface KartuKeluargaUpdate {
  nomorKK?: string;
  nomorRumah?: string;
  namaKepalaKeluarga?: string;
}

export interface Warga {
  id: number;
  kkId: number;
  namaLengkap: string;
  nik: string;
  /** @nullable */
  tempatLahir?: string | null;
  /** @nullable */
  tanggalLahir?: string | null;
  /** @nullable */
  namaOrangTua?: string | null;
  nomorRumah: string;
  /** @nullable */
  nomorTelepon?: string | null;
  /** @nullable */
  clerkUserId?: string | null;
  isActive?: boolean;
  createdAt: string;
  /** @nullable */
  kkNomor?: string | null;
}

export interface WargaPublic {
  id: number;
  namaLengkap: string;
  nomorRumah: string;
  /** @nullable */
  clerkUserId?: string | null;
}

export interface WargaInput {
  kkId: number;
  namaLengkap: string;
  nik: string;
  tempatLahir?: string;
  tanggalLahir?: string;
  namaOrangTua?: string;
  nomorRumah: string;
  nomorTelepon?: string;
}

export interface WargaUpdate {
  kkId?: number;
  namaLengkap?: string;
  nik?: string;
  tempatLahir?: string;
  tanggalLahir?: string;
  namaOrangTua?: string;
  nomorRumah?: string;
  nomorTelepon?: string;
  isActive?: boolean;
}

export interface LinkWargaUser {
  wargaId: number;
  clerkUserId: string;
}

export type IuranStatus = (typeof IuranStatus)[keyof typeof IuranStatus];

export const IuranStatus = {
  belum_bayar: "belum_bayar",
  menunggu_verifikasi: "menunggu_verifikasi",
  lunas: "lunas",
} as const;

export interface Iuran {
  id: number;
  wargaId: number;
  bulan: number;
  tahun: number;
  jumlah: number;
  status: IuranStatus;
  /** @nullable */
  buktiBayarUrl?: string | null;
  createdAt: string;
  /** @nullable */
  verifiedAt?: string | null;
  /** @nullable */
  verifiedBy?: string | null;
  /** @nullable */
  namaWarga?: string | null;
  /** @nullable */
  nomorRumah?: string | null;
}

export interface IuranInput {
  wargaId: number;
  bulan: number;
  tahun: number;
  jumlah: number;
}

export type IuranStatusUpdateStatus =
  (typeof IuranStatusUpdateStatus)[keyof typeof IuranStatusUpdateStatus];

export const IuranStatusUpdateStatus = {
  belum_bayar: "belum_bayar",
  menunggu_verifikasi: "menunggu_verifikasi",
  lunas: "lunas",
} as const;

export interface IuranStatusUpdate {
  status: IuranStatusUpdateStatus;
}

export interface BuktiBayarInput {
  buktiBayarUrl: string;
}

export interface GenerateMonthlyIuran {
  bulan: number;
  tahun: number;
  jumlah: number;
}

export interface GenerateMonthlyResult {
  created: number;
  skipped: number;
}

export type SuratStatus = (typeof SuratStatus)[keyof typeof SuratStatus];

export const SuratStatus = {
  diajukan: "diajukan",
  diproses: "diproses",
  selesai: "selesai",
} as const;

export interface Surat {
  id: number;
  /** @nullable */
  clerkUserId?: string | null;
  namaLengkap: string;
  /** @nullable */
  nomorRumah?: string | null;
  jenisSurat: string;
  alasan: string;
  status: SuratStatus;
  /** @nullable */
  catatanAdmin?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface SuratInput {
  namaLengkap: string;
  nomorRumah?: string;
  jenisSurat: string;
  alasan: string;
}

export type SuratStatusUpdateStatus =
  (typeof SuratStatusUpdateStatus)[keyof typeof SuratStatusUpdateStatus];

export const SuratStatusUpdateStatus = {
  diajukan: "diajukan",
  diproses: "diproses",
  selesai: "selesai",
} as const;

export interface SuratStatusUpdate {
  status: SuratStatusUpdateStatus;
  catatanAdmin?: string;
}

export interface Pengumuman {
  id: number;
  judul: string;
  konten: string;
  /** @nullable */
  kategori?: string | null;
  /** @nullable */
  authorId?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface PengumumanInput {
  judul: string;
  konten: string;
  kategori?: string;
}

export interface PengumumanUpdate {
  judul?: string;
  konten?: string;
  kategori?: string;
}

export interface DashboardSummary {
  totalWarga: number;
  totalKK: number;
  iuranLunasBulanIni: number;
  iuranBelumBayar: number;
  iuranMenungguVerifikasi: number;
  suratBaru: number;
  totalPemasukan: number;
}

export interface PendingActions {
  buktiPembayaran: number;
  permohonanSurat: number;
}

export type ListKKParams = {
  search?: string;
};

export type ListWargaParams = {
  search?: string;
  kkId?: number;
};

export type ListIuranParams = {
  wargaId?: number;
  status?: ListIuranStatus;
  tahun?: number;
  bulan?: number;
};

export type ListIuranStatus =
  (typeof ListIuranStatus)[keyof typeof ListIuranStatus];

export const ListIuranStatus = {
  belum_bayar: "belum_bayar",
  menunggu_verifikasi: "menunggu_verifikasi",
  lunas: "lunas",
} as const;

export type ListSuratParams = {
  status?: ListSuratStatus;
};

export type ListSuratStatus =
  (typeof ListSuratStatus)[keyof typeof ListSuratStatus];

export const ListSuratStatus = {
  diajukan: "diajukan",
  diproses: "diproses",
  selesai: "selesai",
} as const;