# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RT Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /kk:
    get:
      operationId: listKK
      tags: [kk]
      summary: List all family cards
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of family cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KartuKeluarga"
    post:
      operationId: createKK
      tags: [kk]
      summary: Create a new family card
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KartuKeluargaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KartuKeluarga"
  /kk/{id}:
    get:
      operationId: getKK
      tags: [kk]
      summary: Get a family card
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Family card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KartuKeluarga"
    patch:
      operationId: updateKK
      tags: [kk]
      summary: Update a family card
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KartuKeluargaUpdate"
      responses:
        "200":
          description: Updated family card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KartuKeluarga"
    delete:
      operationId: deleteKK
      tags: [kk]
      summary: Delete a family card
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ─── Warga ────────────────────────────────────────────────────────────────
...
  /iuran:
    get:
      operationId: listIuran
      tags: [iuran]
      summary: List dues records
      parameters:
        - in: query
          name: wargaId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [belum_bayar, menunggu_verifikasi, lunas]
        - in: query
          name: tahun
          schema:
            type: integer
        - in: query
          name: bulan
          schema:
            type: integer
      responses:
        "200":
          description: List of dues records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Iuran"
    post:
      operationId: createIuran
      tags: [iuran]
      summary: Create a dues record (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IuranInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Iuran"
  /iuran/generate-monthly:
    post:
      operationId: generateMonthlyIuran
      tags: [iuran]
      summary: Generate monthly dues for all active residents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateMonthlyIuran"
      responses:
        "200":
          description: Generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateMonthlyResult"
...
  /iuran/{id}:
    get:
      operationId: getIuran
      tags: [iuran]
      summary: Get a dues record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dues record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Iuran"
    patch:
      operationId: updateIuranStatus
      tags: [iuran]
      summary: Update dues status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IuranStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Iuran"
  /iuran/{id}/bukti-bayar:
    post:
      operationId: uploadBuktiBayar
      tags: [iuran]
      summary: Upload payment proof URL
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuktiBayarInput"
      responses:
        "200":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Iuran"

  # ─── Surat / Letter Requests ──────────────────────────────────────────────
...
  /pengumuman:
    get:
      operationId: listPengumuman
      tags: [pengumuman]
      summary: List announcements
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pengumuman"
    post:
      operationId: createPengumuman
      tags: [pengumuman]
      summary: Create an announcement (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PengumumanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pengumuman"
  /pengumuman/{id}:
    patch:
      operationId: updatePengumuman
      tags: [pengumuman]
      summary: Update an announcement
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PengumumanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pengumuman"
    delete:
      operationId: deletePengumuman
      tags: [pengumuman]
      summary: Delete an announcement (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # ─── Dashboard ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    KartuKeluarga:
      type: object
      required: [id, nomorKK, nomorRumah, namaKepalaKeluarga, createdAt]
      properties:
        id:
          type: integer
        nomorKK:
          type: string
        nomorRumah:
          type: string
        namaKepalaKeluarga:
          type: string
        createdAt:
          type: string
        wargaCount:
          type: ["integer", "null"]
    KartuKeluargaInput:
      type: object
      required: [nomorKK, nomorRumah, namaKepalaKeluarga]
      properties:
        nomorKK:
          type: string
        nomorRumah:
          type: string
        namaKepalaKeluarga:
          type: string
    KartuKeluargaUpdate:
      type: object
      properties:
        nomorKK:
          type: string
        nomorRumah:
          type: string
        namaKepalaKeluarga:
          type: string

    # ─── Warga ───────────────────────────────────────────────────────────────
...
    Iuran:
      type: object
      required: [id, wargaId, bulan, tahun, jumlah, status, createdAt]
      properties:
        id:
          type: integer
        wargaId:
          type: integer
        bulan:
          type: integer
        tahun:
          type: integer
        jumlah:
          type: integer
        status:
          type: string
          enum: [belum_bayar, menunggu_verifikasi, lunas]
        buktiBayarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        verifiedAt:
          type: ["string", "null"]
        verifiedBy:
          type: ["string", "null"]
        namaWarga:
          type: ["string", "null"]
        nomorRumah:
          type: ["string", "null"]
    IuranInput:
      type: object
      required: [wargaId, bulan, tahun, jumlah]
      properties:
        wargaId:
          type: integer
        bulan:
          type: integer
        tahun:
          type: integer
        jumlah:
          type: integer
    IuranStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [belum_bayar, menunggu_verifikasi, lunas]
    BuktiBayarInput:
      type: object
      required: [buktiBayarUrl]
      properties:
        buktiBayarUrl:
          type: string
    GenerateMonthlyIuran:
      type: object
      required: [bulan, tahun, jumlah]
      properties:
        bulan:
          type: integer
        tahun:
          type: integer
        jumlah:
          type: integer
    GenerateMonthlyResult:
      type: object
      required: [created, skipped]
      properties:
        created:
          type: integer
        skipped:
          type: integer

    # ─── Surat ───────────────────────────────────────────────────────────────
...
    Pengumuman:
      type: object
      required: [id, judul, konten, createdAt]
      properties:
        id:
          type: integer
        judul:
          type: string
        konten:
          type: string
        kategori:
          type: ["string", "null"]
        authorId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    PengumumanInput:
      type: object
      required: [judul, konten]
      properties:
        judul:
          type: string
        konten:
          type: string
        kategori:
          type: string
    PengumumanUpdate:
      type: object
      properties:
        judul:
          type: string
        konten:
          type: string
        kategori:
          type: string

    # ─── Dashboard ────────────────────────────────────────────────────────────
...