# lib/api-zod/src/generated/types/iuran.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RT Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { IuranStatus } from "./iuranStatus";

export interface Iuran {
  id: number;
  wargaId: number;
  bulan: number;
  tahun: number;
  jumlah: number;
  status: IuranStatus;
  /** @nullable */
  buktiBayarUrl?: string | null;
  createdAt: string;
  /** @nullable */
  verifiedAt?: string | null;
  /** @nullable */
  verifiedBy?: string | null;
  /** @nullable */
  namaWarga?: string | null;
  /** @nullable */
  nomorRumah?: string | null;
}