# lib/api-zod/src/generated/types/iuranInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RT Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface IuranInput {
  wargaId: number;
  bulan: number;
  tahun: number;
  jumlah: number;
}