# lib/api-zod/src/generated/types/pengumumanInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RT Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface PengumumanInput {
  judul: string;
  konten: string;
  kategori?: string;
}